dashboard/requirements.txt
redis-cli==5.2.1
dash==2.18.2